refactor: consolidate FakeChannel implementation into wrapper_contract_test_utils.hpp and update tests to use the new class
